refactor(maven): split pom parser state by domain
